add teacher update,delete and getstudent services

+++ src/components/StudentSectionList.jsx
@@ -18,12 +18,16 @@ import {
 } from "@mui/material";
 import { useMediaQuery, useTheme } from "@mui/material";
 import {
+  adminDeleteStudent,
+  adminGetStudentList,
+  adminUpdateStudent,
+  deleteStudent,
   getStudentList,
   updateStudent,
-  deleteStudent,
 } from "../services/Axios.service";
 
 function StudentSectionList({ sectionId }) {
+  const role = useSelector((state) => state.appAuth.role);
   const theme = useTheme();
   const isSmallScreen = useMediaQuery(theme.breakpoints.down("md"));
 
@@ -52,8 +56,12 @@ function StudentSectionList({ sectionId }) {
   const [idForDelete, setIdForDelete] = useState();
 
   const getStudent = async () => {
-    const studentList = await getStudentList(sectionId, pageNo);
-    // console.log(studentList);
+    let studentList;
+    if (role === "teacher") {
+      studentList = await getStudentList(sectionId, pageNo);
+    } else {
+      studentList = await adminGetStudentList(sectionId, pageNo);
+    }
     setTotalStudentCount(studentList.data.result.totalCount);
     setLimit(studentList.data.result.limit);
     setStudentData(studentList.data.result.studentList);
@@ -117,7 +125,13 @@ function StudentSectionList({ sectionId }) {
 
   const handleConfirmDelete = async () => {
     try {
-      await deleteStudent(idForDelete);
+      if (role === "teacher") {
+        await deleteStudent(idForDelete);
+        console.log("t");
+      } else {
+        await adminDeleteStudent(idForDelete);
+        console.log("a");
+      }
       getStudent();
       toast.success("Student deleted successfully!");
     } catch (error) {
@@ -151,7 +165,11 @@ function StudentSectionList({ sectionId }) {
       ...formValues,
     };
     try {
-      await updateStudent(updatedStudent["_id"], updatedStudent);
+      if (role === "teacher") {
+        await updateStudent(updatedStudent["_id"], updatedStudent);
+      } else {
+        await adminUpdateStudent(updatedStudent["_id"], updatedStudent);
+      }
       getStudent();
       toast.success("Student updated successfully!");
     } catch (error) {
@@ -253,11 +271,11 @@ function StudentSectionList({ sectionId }) {
                     {pageNo * limit - (limit - 1)} -{" "}
                     {Math.min(totalStudentCount, pageNo * limit)}
                   </span>
-                  <span className="leading-5 text-slate-400">from</span>{" "}
+                  <span className="leading-5 text-slate-400">{' '}from</span>{" "}
                   <span className="leading-5 text-blue-950">
                     {totalStudentCount}
                   </span>
-                  <span className="leading-5 text-slate-400">data</span>
+                  <span className="leading-5 text-slate-400">{' '}data</span>
                 </div>
                 <Stack spacing={2}>
                   <Pagination
